move headers to inc/_internal
