+++ sdk/core/core/inc/_internal/az_curl_adapter.h
@@ -4,8 +4,8 @@
 #ifndef _az_CURL_ADAPTER_H
 #define _az_CURL_ADAPTER_H
 
-#include <az_http_request_builder.h>
-#include <az_http_response.h>
+#include <_internal/az_http_request_builder.h>
+#include <_internal/az_http_response.h>
 #include <az_result.h>
 
 #include <curl/curl.h>
