+++ sdk/core/core/inc/_internal/az_curl_slist.h
