+++ sdk/core/core/inc/_internal/az_hex.h
