+++ sdk/core/core/inc/_internal/az_http_pipeline.h
@@ -4,9 +4,9 @@
 #ifndef _az_HTTP_PIPELINE_H
 #define _az_HTTP_PIPELINE_H
 
-#include <az_http_policy.h>
-#include <az_http_request_builder.h>
-#include <az_http_response.h>
+#include <_internal/az_http_policy.h>
+#include <_internal/az_http_request_builder.h>
+#include <_internal/az_http_response.h>
 #include <az_result.h>
 
 #include <_az_cfg_prefix.h>
