+++ sdk/core/core/inc/_internal/az_http_response_parser.h
@@ -4,10 +4,10 @@
 #ifndef _az_HTTP_RESPONSE_PARSER_H
 #define _az_HTTP_RESPONSE_PARSER_H
 
+#include <_internal/az_span_reader.h>
 #include <az_pair.h>
 #include <az_result.h>
 #include <az_span.h>
-#include <az_span_reader.h>
 
 #include <stdint.h>
