+++ sdk/core/core/inc/_internal/az_span_malloc.h
