+++ sdk/core/core/inc/_internal/az_span_reader.h
