+++ sdk/core/core/src/az_curl_slist.c
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_curl_slist.h>
+#include <_internal/az_curl_slist.h>
 
 #include <az_action.h>
-#include <az_http_header.h>
+#include <_internal/az_http_header.h>
 
 #include <_az_cfg.h>
