+++ sdk/core/core/src/az_http_query.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_http_query.h>
+#include <_internal/az_http_query.h>
 
 #include <az_str.h>
 
@@ -36,7 +36,8 @@ AZ_ACTION_FUNC(az_http_query_param, az_http_query_state, az_pair_action)
  *
  * Note: currently, the function assumes that pair.key and pair.value encoded into URL format.
  */
-AZ_NODISCARD az_result az_http_query_param(az_http_query_state * const p_state, az_pair const query_param) {
+AZ_NODISCARD az_result
+az_http_query_param(az_http_query_state * const p_state, az_pair const query_param) {
   AZ_CONTRACT_ARG_NOT_NULL(p_state);
 
   az_span_action const write_span = p_state->write_span;
