+++ sdk/core/core/src/az_http_request_builder.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_http_request_builder.h>
+#include <_internal/az_http_request_builder.h>
 
 #include <az_contract.h>
 #include <az_str.h>
