+++ sdk/core/core/src/az_json_string.c
@@ -3,7 +3,7 @@
 
 #include <az_json_string.h>
 
-#include <az_hex.h>
+#include <_internal/az_hex.h>
 #include <az_str.h>
 
 #include <ctype.h>
