+++ sdk/core/core/src/az_span_reader.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_span_reader.h>
+#include <_internal/az_span_reader.h>
 
 #include <ctype.h>
