+++ sdk/core/core/src/az_uri.c
@@ -1,10 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_uri.h>
+#include <_internal/az_uri.h>
 
+#include <_internal/az_hex.h>
 #include <az_contract.h>
-#include <az_hex.h>
 
 #include <assert.h>
 #include <ctype.h>
