+++ sdk/core/core/test/test_json_pointer.c
@@ -3,7 +3,7 @@
 
 #include <az_json_pointer.h>
 #include <az_str.h>
-#include <az_span_reader.h>
+#include <_internal/az_span_reader.h>
 
 #include "./az_test.h"
