+++ sdk/identity/identity/inc/az_identity_credential.h
@@ -5,7 +5,7 @@
 #define _az_IDENTITY_CREDENTIAL_H
 
 #include <az_contract.h>
-#include <az_http_request_builder.h>
+#include <_internal/az_http_request_builder.h>
 #include <az_result.h>
 
 #include <_az_cfg_prefix.h>
