+++ sdk/identity/identity/test/curl_easy_perform_poc.c
@@ -1,15 +1,15 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_http_pipeline.h>
-#include <az_http_request_builder.h>
-#include <az_http_response_parser.h>
+#include <_internal/az_http_pipeline.h>
+#include <_internal/az_http_request_builder.h>
+#include <_internal/az_http_response_parser.h>
+#include <_internal/az_span_malloc.h>
 #include <az_identity_access_token_context.h>
 #include <az_identity_client_secret_credential.h>
 #include <az_pair.h>
 #include <az_span.h>
 #include <az_span_builder.h>
-#include <az_span_malloc.h>
 #include <az_str.h>
 
 #include <stdio.h>
