+++ sdk/storage/blobs/src/az_storage_blobs_blob_client.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_http_header.h>
+#include <_internal/az_http_header.h>
 #include <az_json_builder.h>
 #include <az_storage_blobs.h>
