+++ sdk/storage/blobs/test/storage_blobs_poc.c
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <az_http_response_parser.h>
+#include <_internal/az_http_response_parser.h>
 #include <az_identity_client_secret_credential.h>
 #include <az_json_get.h>
 #include <az_storage_blobs.h>
